--- kolena/_experimental/dataset/__init__.py
@@ -1,27 +1,0 @@
-# Copyright 2021-2024 Kolena Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# noreorder
-from kolena._experimental.dataset._dataset import fetch_dataset_history
-from kolena._experimental.dataset._dataset import fetch_dataset
-from kolena._experimental.dataset._dataset import register_dataset
-from kolena._experimental.dataset._evaluation import fetch_results
-from kolena._experimental.dataset._evaluation import test
-
-__all__ = [
-    "register_dataset",
-    "fetch_dataset_history",
-    "fetch_dataset",
-    "fetch_results",
-    "test",
-]
