+++ kolena/dataset/evaluation.py
@@ -26,16 +26,6 @@ import pandas as pd
 from kolena._api.v2.model import LoadResultsRequest
 from kolena._api.v2.model import Path
 from kolena._api.v2.model import UploadResultsRequest
-from kolena._experimental.dataset._dataset import _iter_dataset_raw
-from kolena._experimental.dataset._dataset import _to_deserialized_dataframe
-from kolena._experimental.dataset._dataset import _to_serialized_dataframe
-from kolena._experimental.dataset._dataset import load_dataset
-from kolena._experimental.dataset.common import COL_DATAPOINT
-from kolena._experimental.dataset.common import COL_DATAPOINT_ID_OBJECT
-from kolena._experimental.dataset.common import COL_EVAL_CONFIG
-from kolena._experimental.dataset.common import COL_RESULT
-from kolena._experimental.dataset.common import validate_batch_size
-from kolena._experimental.dataset.common import validate_dataframe_ids
 from kolena._utils import krequests_v2 as krequests
 from kolena._utils import log
 from kolena._utils.batched_load import _BatchedLoader
@@ -43,6 +33,16 @@ from kolena._utils.batched_load import init_upload
 from kolena._utils.batched_load import upload_data_frame
 from kolena._utils.consts import BatchSize
 from kolena._utils.state import API_V2
+from kolena.dataset.common import COL_DATAPOINT
+from kolena.dataset.common import COL_DATAPOINT_ID_OBJECT
+from kolena.dataset.common import COL_EVAL_CONFIG
+from kolena.dataset.common import COL_RESULT
+from kolena.dataset.common import validate_batch_size
+from kolena.dataset.common import validate_dataframe_ids
+from kolena.dataset.dataset import _iter_dataset_raw
+from kolena.dataset.dataset import _to_deserialized_dataframe
+from kolena.dataset.dataset import _to_serialized_dataframe
+from kolena.dataset.dataset import load_dataset
 from kolena.errors import IncorrectUsageError
 from kolena.errors import NotFoundError
 
@@ -50,7 +50,7 @@ TYPE_EVALUATION_CONFIG = Optional[Dict[str, Any]]
 TEST_ON_TYPE = Optional[Union[str, List[str]]]
 
 
-def _fetch_dataset(dataset: str) -> pd.DataFrame:
+def fetch_dataset(dataset: str) -> pd.DataFrame:
     df_data_batch = list(_iter_dataset_raw(dataset))
     df_datapoints = pd.concat(df_data_batch) if df_data_batch else pd.DataFrame(columns=["id", COL_DATAPOINT])
     df_datapoints.rename(columns={"id": "datapoint_id"}, inplace=True)
