+++ tests/unit/_experimental/dataset/test_dataset.py
@@ -22,17 +22,17 @@ from pandas.testing import assert_frame_equal
 from .data import a_text
 from .data import b_text
 from kolena._api.v2.dataset import EntityData
-from kolena._experimental.dataset._dataset import _add_datatype
-from kolena._experimental.dataset._dataset import _flatten_composite
-from kolena._experimental.dataset._dataset import _infer_datatype
-from kolena._experimental.dataset._dataset import _infer_datatype_value
-from kolena._experimental.dataset._dataset import _infer_id_fields
-from kolena._experimental.dataset._dataset import _to_deserialized_dataframe
-from kolena._experimental.dataset._dataset import _to_serialized_dataframe
-from kolena._experimental.dataset._dataset import DatapointType
-from kolena._experimental.dataset._dataset import resolve_id_fields
-from kolena._experimental.dataset.common import COL_DATAPOINT
-from kolena._experimental.dataset.common import COL_RESULT
+from kolena.dataset import COL_DATAPOINT
+from kolena.dataset import COL_RESULT
+from kolena.dataset.dataset import _add_datatype
+from kolena.dataset.dataset import _flatten_composite
+from kolena.dataset.dataset import _infer_datatype
+from kolena.dataset.dataset import _infer_datatype_value
+from kolena.dataset.dataset import _infer_id_fields
+from kolena.dataset.dataset import _to_deserialized_dataframe
+from kolena.dataset.dataset import _to_serialized_dataframe
+from kolena.dataset.dataset import DatapointType
+from kolena.dataset.dataset import resolve_id_fields
 from kolena.errors import InputValidationError
 from kolena.workflow._datatypes import DATA_TYPE_FIELD
 from kolena.workflow.annotation import BoundingBox
